+++ lib/laams_push.dart
@@ -1,5 +1,6 @@
 library laams_push;
 
+export 'src/adapters/laams_route_mapper.dart';
 export 'src/adapters/laams_route_parser.dart';
 export 'src/adapters/route_authenticator.dart';
 export 'src/entities/enums/animation_type.dart';
@@ -7,7 +8,5 @@ export 'src/entities/laams_route.dart';
 export 'src/laams_page.dart';
 export 'src/laams_push.dart';
 export 'src/laams_push_app.dart';
-export 'src/laams_push_app.dart';
 export 'src/laams_router_delegate.dart';
-// export 'src/state/laams_push.dart';
-export 'src/state/laams_push_provider.dart';
+export 'src/state/laams_push_state.dart';