+++ lib/src/laams_router_delegate.dart
@@ -3,11 +3,11 @@ import 'package:flutter/material.dart';
 
 import 'entities/laams_route.dart';
 import 'laams_page.dart';
-import 'laams_push.dart';
+import 'state/laams_push_state.dart';
 
-class LaamsRouterDelegate extends RouterDelegate<LaamsRoutes>
-    with ChangeNotifier, PopNavigatorRouterDelegateMixin<LaamsRoutes> {
-  final LaamsPush laamsPush;
+class LaamsRouterDelegate extends RouterDelegate<LaamsRoute>
+    with ChangeNotifier, PopNavigatorRouterDelegateMixin<LaamsRoute> {
+  final LaamsPushState laamsPush;
   final LaamsPage Function(LaamsRoute name) onGeneratePages;
   LaamsRouterDelegate(this.laamsPush, this.onGeneratePages)
       : navigatorKey = GlobalKey<NavigatorState>(),
@@ -19,20 +19,18 @@ class LaamsRouterDelegate extends RouterDelegate<LaamsRoutes>
   final GlobalKey<NavigatorState> navigatorKey;
 
   @override
-  List<LaamsRoute> get currentConfiguration => laamsPush.routes;
+  LaamsRoute get currentConfiguration => laamsPush.routes.last;
 
   @override
-  Future<void> setNewRoutePath(LaamsRoutes configuration) {
-    laamsPush.setRoutes(configuration, true);
+  Future<void> setNewRoutePath(LaamsRoute configuration) {
+    // if (kDebugMode) print('LaamsRouterDelegate.setNewRoutePath:$configuration');
+    laamsPush.onResetRoutes(configuration, isInitial: true);
     return SynchronousFuture<void>(null);
   }
 
   bool _onPopPage(Route<dynamic> route, dynamic result) {
     final bool success = route.didPop(result);
-    if (success && laamsPush.routes.length > 1) {
-      laamsPush.pop();
-      notifyListeners();
-    }
+    if (success) laamsPush.onPopRoute();
     return success;
   }