+++ CHANGELOG.md
@@ -1,10 +1,19 @@
-# 0.0.2
+# 1.0.0
+* Released the stable version
+* It has all the listed functionalities
+* Added helper methods for ease of use. 
 
+
+# 0.1.0
+* Added more APIs and documentation, 
+* The router works without authentication
+  
+
+# 0.0.2
 * Minor changes in documention. 
   
 
 # 0.0.1
-
 Initial Version of the library.
 
 * A simple api for using the laams_push